requirements.txt
Django==4.2.28
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
django-filter==24.3
python-dotenv==1.0.1